Extract file upload logic
